pyproject.toml
[project]
name = "aero-ref"
version = "1.0.0"
description = "Aero reference stack — FastMCP flight (FlightAware AeroAPI) + BigQuery tools, LangGraph / Ollama friendly"
readme = "README.md"
requires-python = ">=3.11"
 
dependencies = [
    "fastmcp>=2.0.0",
    "mcp-use>=0.1.0",
    "langchain-ollama>=0.2.0",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "google-cloud-bigquery>=3.25.0",
]
 
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aero_ref"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]